requirements.txt
pandas
python-dotenv
llama-index
llama-index-experimental
llama-index-embeddings-huggingface
llama-index-readers-file
llama-index-vector-stores-postgres
llama-index-llms-llama-cpp 
pymupdf
openpyxl
psycopg2-binary
pgvector
asyncpg
"sqlalchemy[asyncio]"
greenlet
docx2txt